moved some stuff that should have been in board component from topic list
