+++ frontend/src/app/board/board.component.html
@@ -1,5 +1,7 @@
 <app-top-bar></app-top-bar>
 <div class="content">
     <a [href]="createTopicLink" class="btn btn-primary">Create Topic</a>
-    <app-topic-list></app-topic-list>
+    <h1>{{boardTitle}}</h1>
+    <p>{{description}}</p>
+    <app-topic-list [topics]="topics"></app-topic-list>
 </div>
