+++ frontend/src/app/topic-list/topic-list.component.html
@@ -1,6 +1,3 @@
-<h1>{{boardTitle}}</h1>
-<p>{{description}}</p>
-
 <div>
     <div *ngFor="let topic of topics" class="col-sm-12 topic">
         <div class="col-sm-2 left-stats">
